// lib/bambu/profile-config-bounds.generated.ts
export type ProfileConfigDef = {
  type: string;
  min?: number;
  max?: number;
  enumValues?: readonly string[];
};